Improved day 10

+++ src/y2024/day10.rs
@@ -2,7 +2,7 @@ use std::collections::VecDeque;
 
 use aoc24::Coordinate;
 
-fn part1(start: Coordinate<usize>, map: &Vec<Vec<i32>>) -> u32 {
+fn part1(start: Coordinate<usize>, map: &Vec<Vec<u8>>) -> u32 {
     let max = map.len();
     let vertical = Coordinate::new(1, 0);
     let horizontal = Coordinate::new(0, 1);
@@ -22,22 +22,22 @@ fn part1(start: Coordinate<usize>, map: &Vec<Vec<i32>>) -> u32 {
             continue;
         }
         let u = o - vertical;
-        if !u.out_of_bounds(max as usize) && !visited.iter().any(|k| *k == u) && map[u.x][u.y] - val == 1 {
+        if !u.out_of_bounds(max) && !visited.iter().any(|k| *k == u) && map[u.x][u.y] - val == 1 {
             visited.push(u);
             open.push_back(u);
         }
         let d = o + vertical;
-        if !d.out_of_bounds(max as usize) && !visited.iter().any(|k| *k == d) && map[d.x][d.y] - val == 1 {
+        if !d.out_of_bounds(max) && !visited.iter().any(|k| *k == d) && map[d.x][d.y] - val == 1 {
             visited.push(d);
             open.push_back(d);
         }
         let r = o + horizontal;
-        if !r.out_of_bounds(max as usize) && !visited.iter().any(|k| *k == r) && map[r.x][r.y] - val == 1 {
+        if !r.out_of_bounds(max) && !visited.iter().any(|k| *k == r) && map[r.x][r.y] - val == 1 {
             visited.push(r);
             open.push_back(r);
         }
         let l = o - horizontal;
-        if !l.out_of_bounds(max as usize) && !visited.iter().any(|k| *k == l) && map[l.x][l.y] - val == 1 {
+        if !l.out_of_bounds(max) && !visited.iter().any(|k| *k == l) && map[l.x][l.y] - val == 1 {
             visited.push(l);
             open.push_back(l);
         }
@@ -45,7 +45,7 @@ fn part1(start: Coordinate<usize>, map: &Vec<Vec<i32>>) -> u32 {
     count
 }
 
-fn part2(start: Coordinate<usize>, mapa: &Vec<Vec<i32>>) -> u32 {
+fn part2(start: Coordinate<usize>, mapa: &Vec<Vec<u8>>) -> u32 {
     let max = mapa.len();
     let vertical = Coordinate::new(1, 0);
     let horizontal = Coordinate::new(0, 1);
@@ -63,19 +63,19 @@ fn part2(start: Coordinate<usize>, mapa: &Vec<Vec<i32>>) -> u32 {
             continue;
         }
         let u = o - vertical;
-        if !u.out_of_bounds(max as usize) && mapa[u.x][u.y] - val == 1 {
+        if !u.out_of_bounds(max) && mapa[u.x][u.y] - val == 1 {
             open.push_back(u);
         }
         let d = o + vertical;
-        if !d.out_of_bounds(max as usize) && mapa[d.x][d.y] - val == 1 {
+        if !d.out_of_bounds(max) && mapa[d.x][d.y] - val == 1 {
             open.push_back(d);
         }
         let r = o + horizontal;
-        if !r.out_of_bounds(max as usize) && mapa[r.x][r.y] - val == 1 {
+        if !r.out_of_bounds(max) && mapa[r.x][r.y] - val == 1 {
             open.push_back(r);
         }
         let l = o - horizontal;
-        if !l.out_of_bounds(max as usize) && mapa[l.x][l.y] - val == 1 {
+        if !l.out_of_bounds(max) && mapa[l.x][l.y] - val == 1 {
             open.push_back(l);
         }
     }
@@ -84,14 +84,14 @@ fn part2(start: Coordinate<usize>, mapa: &Vec<Vec<i32>>) -> u32 {
 
 pub fn run(inp: Vec<String>) -> (i64, i64) {
     let mut head: Vec<Coordinate<usize>> = vec![];
-    let map: Vec<Vec<i32>> = inp
+    let map: Vec<Vec<u8>> = inp
         .iter()
         .enumerate()
         .map(|(i, f)| {
             f.chars()
                 .enumerate()
                 .map(|(j, c)| {
-                    let n = c.to_digit(10).unwrap() as i32;
+                    let n = c.to_digit(10).unwrap() as u8;
                     if n == 0 {
                         head.push(Coordinate::new(i, j));
                     }
